+++ client/src/Screen/03_ThirdTab/Designer/Branding/createBrandingScreen.js
@@ -1,13 +1,15 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, createRef } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { StyleSheet, Text, View, ScrollView, Image, TouchableOpacity, Alert } from 'react-native';
 
-import BrandingAPI from '../../../Api/branding';
-
-import Loading from '../../../Components/Common/loading';
 import Icon from 'react-native-vector-icons/Ionicons';
-import BrandingInput from '../../../Components/Branding/brandingInput';
-import { getBrandingListByDesignerId } from '../../../Contexts/Designer/Branding';
+import BrandingAPI from '../../../../Api/branding';
+
+import Line from '../../../../Components/Common/line';
+import Loading from '../../../../Components/Common/loading';
+import BrandingInput from '../../../../Components/Branding/brandingInput';
+import BrandingRefInput from '../../../../Components/Branding/brandingRefInput';
+import { getBrandingListByDesignerId } from '../../../../Contexts/Designer/Branding';
 
 function CreateBrandingScreen({ navigation }) {
   const { data: userInfo } = useSelector((state) => state.user);
@@ -23,9 +25,9 @@ function CreateBrandingScreen({ navigation }) {
   const [styleList, setStyleList] = useState('');
 
   const addStyleTags = () => {
-    if (styleTags.length > 4) {
+    if (styleTags.length > 2) {
       setTagText('');
-      return Alert.alert('스타일 태그는 최대 5개까지 입력이 가능합니다');
+      return Alert.alert('스타일 태그는 최대 3개까지 입력이 가능합니다');
     }
     setStyleTags([...styleTags, tagText]);
     setTagText('');
@@ -77,28 +79,6 @@ function CreateBrandingScreen({ navigation }) {
   return (
     <ScrollView style={styles.container}>
       <BrandingInput
-        title="헤어샵"
-        value={shopName}
-        setValue={setShopName}
-        placeholderMessage="근무중인 헤어샵을 입력해주세요"
-        placeholderColor="#878787"
-      />
-      <BrandingInput
-        title="헤어샵 위치"
-        value={shopAddress}
-        setValue={setShopAddress}
-        placeholderMessage="서울특별시 강남구"
-        placeholderColor="#878787"
-      />
-      <BrandingInput
-        title="직급"
-        value={position}
-        setValue={setPosition}
-        placeholderMessage="직급을 입력해주세요"
-        placeholderColor="#878787"
-      />
-      <View style={styles.line}></View>
-      <BrandingInput
         title="포트폴리오 이름"
         value={brandingName}
         setValue={setBrandingName}
@@ -106,8 +86,8 @@ function CreateBrandingScreen({ navigation }) {
         placeholderColor="#FF533A"
       />
       <BrandingInput
-        title="스타일 태그"
-        subTitle="(최대 5개)"
+        title="해시 태그"
+        subTitle="(최대 3개)"
         value={tagText}
         setValue={setTagText}
         setStyleTags={setStyleTags}
@@ -132,33 +112,28 @@ function CreateBrandingScreen({ navigation }) {
           ( {description.length} / 400 )
         </Text>
       </View>
-      <View style={styles.inputBox}>
-        <View style={styles.titleTextArea}>
-          <Text style={styles.titleText}>추천 스타일</Text>
-        </View>
-        <View style={styles.styleBox}>
-          <TouchableOpacity style={styles.styleArea}>
-            <Image
-              source={{
-                uri: 'https://bidi-s3.s3.ap-northeast-2.amazonaws.com/test/style_img7.png',
-              }}
-              style={styles.styleImg}
-            />
-          </TouchableOpacity>
-          <TouchableOpacity style={styles.styleArea}>
-            <Image
-              source={{
-                uri: 'https://bidi-s3.s3.ap-northeast-2.amazonaws.com/test/style_img8.png',
-              }}
-              style={styles.styleImg}
-            />
-          </TouchableOpacity>
-          <TouchableOpacity style={[styles.styleArea, styles.addStyleArea]}>
-            <Icon name="add" size={50} style={styles.addIcon} />
-          </TouchableOpacity>
-          <TouchableOpacity style={styles.styleArea}></TouchableOpacity>
-        </View>
-      </View>
+      <BrandingRefInput title="스타일 메뉴" />
+      <BrandingInput
+        title="헤어샵"
+        value={shopName}
+        setValue={setShopName}
+        placeholderMessage="근무중인 헤어샵을 입력해주세요"
+        placeholderColor="#878787"
+      />
+      <BrandingInput
+        title="헤어샵 위치"
+        value={shopAddress}
+        setValue={setShopAddress}
+        placeholderMessage="서울특별시 강남구"
+        placeholderColor="#878787"
+      />
+      <BrandingInput
+        title="직급"
+        value={position}
+        setValue={setPosition}
+        placeholderMessage="직급을 입력해주세요"
+        placeholderColor="#878787"
+      />
       <TouchableOpacity style={styles.registerBtnArea} onPress={registerHandler}>
         <Text style={styles.registerBtnText}>저장하기</Text>
       </TouchableOpacity>
@@ -168,48 +143,7 @@ function CreateBrandingScreen({ navigation }) {
 
 const styles = StyleSheet.create({
   container: { flex: 1 },
-  inputBox: {
-    margin: 16,
-  },
-  titleText: {
-    color: '#111111',
-    fontWeight: 'bold',
-    lineHeight: 20,
-    letterSpacing: -0.5,
-    fontSize: 17,
-  },
-  titleTextArea: {
-    marginBottom: 16,
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
 
-  styleBox: {
-    flexWrap: 'wrap',
-    flexDirection: 'row',
-  },
-  styleArea: {
-    width: 168,
-    height: 168,
-    backgroundColor: '#F5F5F5',
-    borderRadius: 3,
-    marginRight: 8,
-    marginBottom: 8,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  addStyleArea: {
-    backgroundColor: '#DDDDDD',
-  },
-  styleImg: {
-    width: 168,
-    height: 168,
-    resizeMode: 'cover',
-    borderRadius: 3,
-  },
-  addIcon: {
-    color: '#fff',
-  },
   line: {
     height: 10,
     backgroundColor: '#f4f4f4',
