+++ client/src/Screen/03_ThirdTab/Designer/StyleBook/createStyleBookScreen.js
@@ -2,42 +2,40 @@ import React, { useState, useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { StyleSheet, Text, View, ScrollView, TouchableOpacity, Image, Alert } from 'react-native';
 
-import StyleAPI from '../../../Api/style';
-import { createStyleForm } from '../../../Lib/utils';
-import { getStyleListByDesignerId } from '../../../Contexts/Designer/Style';
-import { LENGTH_TYPE, STYLE_TYPE, GENDER_TYPE } from '../../../Lib/constant';
+import StyleAPI from '../../../../Api/style';
+import { createStyleForm } from '../../../../Lib/utils';
+import { getStyleListByDesignerId } from '../../../../Contexts/Designer/Style';
+import { LENGTH_TYPE, STYLE_TYPE, GENDER_TYPE } from '../../../../Lib/constant';
 
-import Line from '../../../Components/Common/line';
-import StyleTags from '../../../Components/StyleBook/styleTags';
-import StyleMain from '../../../Components/StyleBook/styleMain';
-import StyleImage from '../../../Components/StyleBook/styleImage';
-import StyleTitle from '../../../Components/StyleBook/styleTitle';
-import StylePrice from '../../../Components/StyleBook/stylePrice';
-import StyleCategory from '../../../Components/StyleBook/styleCategory';
-import StyleDescription from '../../../Components/StyleBook/styleDescription';
+import Line from '../../../../Components/Common/line';
+import StyleTags from '../../../../Components/StyleBook/styleTags';
+import StyleImage from '../../../../Components/StyleBook/styleImage';
+import StyleTitle from '../../../../Components/StyleBook/styleTitle';
+import StylePrice from '../../../../Components/StyleBook/stylePrice';
+import StyleCategory from '../../../../Components/StyleBook/styleCategory';
+import StyleDescription from '../../../../Components/StyleBook/styleDescription';
 
-function EditStyleBookScreen({ navigation, route }) {
-  const { styleItem } = route.params;
+function CreateStyleBookScreen({ navigation }) {
   const { data: userInfo } = useSelector((state) => state.user);
   const dispatch = useDispatch();
 
-  const [frontStyle, setFrontStyle] = useState(styleItem.front_img_src);
-  const [sideStyle, setSideStyle] = useState(styleItem.side_img_src);
-  const [backStyle, setBackStyle] = useState(styleItem.back_img_src);
-  const [styleTitle, setStyleTitle] = useState(styleItem.title);
-  const [styleDescription, setStyleDescription] = useState(styleItem.description);
-  const [stylePrice, setStylePrice] = useState(String(styleItem.price));
+  const [frontStyle, setFrontStyle] = useState({});
+  const [sideStyle, setSideStyle] = useState({});
+  const [backStyle, setBackStyle] = useState({});
+  const [styleTitle, setStyleTitle] = useState('');
+  const [styleDescription, setStyleDescription] = useState('');
+  const [stylePrice, setStylePrice] = useState('');
   const [tagText, setTagText] = useState('');
-  const [styleTags, setStyleTags] = useState(styleItem.keyword_array);
+  const [styleTags, setStyleTags] = useState([]);
 
   const [genderTypeOpen, setGenderTypeOpen] = useState(false);
-  const [genderTypeValue, setGenderTypeValue] = useState(styleItem.gender_type);
+  const [genderTypeValue, setGenderTypeValue] = useState('미선택');
   const [genderTypeItems, setGenderTypeItems] = useState(GENDER_TYPE);
   const [lengthTypeOpen, setLengthTypeOpen] = useState(false);
-  const [lengthTypeValue, setLengthTypeValue] = useState(styleItem.length_type);
+  const [lengthTypeValue, setLengthTypeValue] = useState('미선택');
   const [lengthTypeItems, setLengthTypeItems] = useState(LENGTH_TYPE);
   const [styleTypeOpen, setStyleTypeOpen] = useState(false);
-  const [styleTypeValue, setStyleTypeValue] = useState(styleItem.style_type);
+  const [styleTypeValue, setStyleTypeValue] = useState(null);
   const [styleTypeItems, setStyleTypeItems] = useState([]);
 
   useEffect(() => {
@@ -58,7 +56,7 @@ function EditStyleBookScreen({ navigation, route }) {
     setStyleTags(filteredStyleTags);
   };
 
-  const editHandler = async () => {
+  const registerHandler = async () => {
     if (!frontStyle) {
       return Alert.alert('스타일 사진을 등록해주세요!');
     }
@@ -90,9 +88,9 @@ function EditStyleBookScreen({ navigation, route }) {
       length_type: lengthTypeValue,
       keyword_array: styleTags.toString(),
     });
-    const response = await StyleAPI.patchStyle(styleItem.id, bodyData);
+    const response = await StyleAPI.registerStyle(bodyData);
     if (response) {
-      Alert.alert('스타일북 수정이 완료되었습니다!');
+      Alert.alert('스타일북이 작성되었습니다!');
       dispatch(getStyleListByDesignerId(userInfo.id));
       navigation.navigate('BrandingMain', { screen: 'StyleBook' });
     }
@@ -108,7 +106,7 @@ function EditStyleBookScreen({ navigation, route }) {
           setSideStyle={setSideStyle}
           backStyle={backStyle}
           setBackStyle={setBackStyle}
-          isEdit={true}
+          isEdit={false}
         />
         <Line />
         <StyleTitle
@@ -164,9 +162,9 @@ function EditStyleBookScreen({ navigation, route }) {
           placeholderMessage="0"
           placeholderColor="#878787"
         />
-        <TouchableOpacity onPress={editHandler}>
+        <TouchableOpacity onPress={registerHandler}>
           <View style={styles.bottomBtnArea}>
-            <Text style={styles.editBtnText}>수정하기</Text>
+            <Text style={styles.registerBtnText}>등록하기</Text>
           </View>
         </TouchableOpacity>
       </View>
@@ -186,7 +184,7 @@ const styles = StyleSheet.create({
     alignItems: 'center',
     justifyContent: 'center',
   },
-  editBtnText: {
+  registerBtnText: {
     color: '#fff',
     fontSize: 18,
     fontWeight: 'normal',
@@ -196,4 +194,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default EditStyleBookScreen;
+export default CreateStyleBookScreen;
