+++ client/src/Screen/03_ThirdTab/Designer/StyleBook/styleBookListScreen.js
@@ -2,10 +2,11 @@ import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { StyleSheet, Text, View, Image, TouchableOpacity, ScrollView } from 'react-native';
 
-import Loading from '../../../Components/Common/loading';
 import Ionicons from 'react-native-vector-icons/Ionicons';
+
+import Loading from '../../../../Components/Common/loading';
 import NoStyleBookScreen from './noStyleBookScreen';
-import { getStyleListByDesignerId } from '../../../Contexts/Designer/Style';
+import { getStyleListByDesignerId } from '../../../../Contexts/Designer/Style';
 
 function StyleBookListScreen({ navigation }) {
   const dispatch = useDispatch();
